Brunch adapter is implemented. #402: tests for liquidation of brunch market were passed

+++ src/test/Aave3PriceOracleMock.sol
@@ -43,27 +43,25 @@ contract Aave3PriceOracleMock is IAavePriceOracle {
         return _aavePriceOracle.BASE_CURRENCY();
     }
 
-   /// @notice Returns the base currency unit
-   /// @dev 1 ether for ETH, 1e8 for USD.
-   /// @return Returns the base currency unit.
-    function BASE_CURRENCY_UNIT() external view override returns (uint256) {
+    /// @notice Returns the base currency unit
+    /// @dev 1 ether for ETH, 1e8 for USD.
+    /// @return Returns the base currency unit.
+    function BASE_CURRENCY_UNIT() external view override returns (uint) {
         return _aavePriceOracle.BASE_CURRENCY_UNIT();
     }
 
     /// @notice Returns the asset price in the base currency
     /// @param asset The address of the asset
     /// @return The price of the asset
-    function getAssetPrice(address asset) external view override returns (uint256) {
+    function getAssetPrice(address asset) external view override returns (uint) {
         uint updatedPrice = _prices[asset];
-        return updatedPrice == 0
-            ? _aavePriceOracle.getAssetPrice(asset)
-            : updatedPrice;
+        return updatedPrice == 0 ? _aavePriceOracle.getAssetPrice(asset) : updatedPrice;
     }
 
     /// @notice Returns a list of prices from a list of assets addresses
     /// @param assets The list of assets addresses
     /// @return The prices of the given assets
-    function getAssetsPrices(address[] memory assets) external view override returns (uint256[] memory) {
+    function getAssetsPrices(address[] memory assets) external view override returns (uint[] memory) {
         uint[] memory dest = new uint[](assets.length);
         for (uint i = 0; i < assets.length; ++i) {
             dest[i] = _prices[assets[i]];
